Refactoring to fit into new layout
