--- tools/approval-service/internal/approvalservice/sources/githubevents/testdata/workflow_dispatch_event.json
@@ -1,137 +1,0 @@
-{
-    "ref": "refs/heads/main",
-    "workflow": ".github/workflows/CI.yml",
-    "inputs": {
-      "string_input": "thisismessage",
-      "number_input": "1234",
-      "bool_input": "true"
-    },
-    "repository": {
-      "id": 947657554,
-      "node_id": "R_kgDOOHwbUg",
-      "name": "test-repo",
-      "full_name": "gravitational/test-repo",
-      "private": false,
-      "owner": {
-        "login": "gravitational",
-        "id": 29820606,
-        "node_id": "MDQ6VXNlcjI5ODIwNjA2",
-        "avatar_url": "https://avatars.githubusercontent.com/u/29820606?v=4",
-        "gravatar_id": "",
-        "url": "https://api.github.com/users/gravitational",
-        "html_url": "https://github.com/gravitational",
-        "followers_url": "https://api.github.com/users/gravitational/followers",
-        "following_url": "https://api.github.com/users/gravitational/following{/other_user}",
-        "gists_url": "https://api.github.com/users/gravitational/gists{/gist_id}",
-        "starred_url": "https://api.github.com/users/gravitational/starred{/owner}{/repo}",
-        "subscriptions_url": "https://api.github.com/users/gravitational/subscriptions",
-        "organizations_url": "https://api.github.com/users/gravitational/orgs",
-        "repos_url": "https://api.github.com/users/gravitational/repos",
-        "events_url": "https://api.github.com/users/gravitational/events{/privacy}",
-        "received_events_url": "https://api.github.com/users/gravitational/received_events",
-        "type": "User",
-        "user_view_type": "public",
-        "site_admin": false
-      },
-      "html_url": "https://github.com/gravitational/test-repo",
-      "description": null,
-      "fork": false,
-      "url": "https://api.github.com/repos/gravitational/test-repo",
-      "forks_url": "https://api.github.com/repos/gravitational/test-repo/forks",
-      "keys_url": "https://api.github.com/repos/gravitational/test-repo/keys{/key_id}",
-      "collaborators_url": "https://api.github.com/repos/gravitational/test-repo/collaborators{/collaborator}",
-      "teams_url": "https://api.github.com/repos/gravitational/test-repo/teams",
-      "hooks_url": "https://api.github.com/repos/gravitational/test-repo/hooks",
-      "issue_events_url": "https://api.github.com/repos/gravitational/test-repo/issues/events{/number}",
-      "events_url": "https://api.github.com/repos/gravitational/test-repo/events",
-      "assignees_url": "https://api.github.com/repos/gravitational/test-repo/assignees{/user}",
-      "branches_url": "https://api.github.com/repos/gravitational/test-repo/branches{/branch}",
-      "tags_url": "https://api.github.com/repos/gravitational/test-repo/tags",
-      "blobs_url": "https://api.github.com/repos/gravitational/test-repo/git/blobs{/sha}",
-      "git_tags_url": "https://api.github.com/repos/gravitational/test-repo/git/tags{/sha}",
-      "git_refs_url": "https://api.github.com/repos/gravitational/test-repo/git/refs{/sha}",
-      "trees_url": "https://api.github.com/repos/gravitational/test-repo/git/trees{/sha}",
-      "statuses_url": "https://api.github.com/repos/gravitational/test-repo/statuses/{sha}",
-      "languages_url": "https://api.github.com/repos/gravitational/test-repo/languages",
-      "stargazers_url": "https://api.github.com/repos/gravitational/test-repo/stargazers",
-      "contributors_url": "https://api.github.com/repos/gravitational/test-repo/contributors",
-      "subscribers_url": "https://api.github.com/repos/gravitational/test-repo/subscribers",
-      "subscription_url": "https://api.github.com/repos/gravitational/test-repo/subscription",
-      "commits_url": "https://api.github.com/repos/gravitational/test-repo/commits{/sha}",
-      "git_commits_url": "https://api.github.com/repos/gravitational/test-repo/git/commits{/sha}",
-      "comments_url": "https://api.github.com/repos/gravitational/test-repo/comments{/number}",
-      "issue_comment_url": "https://api.github.com/repos/gravitational/test-repo/issues/comments{/number}",
-      "contents_url": "https://api.github.com/repos/gravitational/test-repo/contents/{+path}",
-      "compare_url": "https://api.github.com/repos/gravitational/test-repo/compare/{base}...{head}",
-      "merges_url": "https://api.github.com/repos/gravitational/test-repo/merges",
-      "archive_url": "https://api.github.com/repos/gravitational/test-repo/{archive_format}{/ref}",
-      "downloads_url": "https://api.github.com/repos/gravitational/test-repo/downloads",
-      "issues_url": "https://api.github.com/repos/gravitational/test-repo/issues{/number}",
-      "pulls_url": "https://api.github.com/repos/gravitational/test-repo/pulls{/number}",
-      "milestones_url": "https://api.github.com/repos/gravitational/test-repo/milestones{/number}",
-      "notifications_url": "https://api.github.com/repos/gravitational/test-repo/notifications{?since,all,participating}",
-      "labels_url": "https://api.github.com/repos/gravitational/test-repo/labels{/name}",
-      "releases_url": "https://api.github.com/repos/gravitational/test-repo/releases{/id}",
-      "deployments_url": "https://api.github.com/repos/gravitational/test-repo/deployments",
-      "created_at": "2025-03-13T03:23:35Z",
-      "updated_at": "2025-05-13T16:30:09Z",
-      "pushed_at": "2025-05-13T16:30:04Z",
-      "git_url": "git://github.com/gravitational/test-repo.git",
-      "ssh_url": "git@github.com:gravitational/test-repo.git",
-      "clone_url": "https://github.com/gravitational/test-repo.git",
-      "svn_url": "https://github.com/gravitational/test-repo",
-      "homepage": null,
-      "size": 8,
-      "stargazers_count": 0,
-      "watchers_count": 0,
-      "language": null,
-      "has_issues": true,
-      "has_projects": true,
-      "has_downloads": true,
-      "has_wiki": true,
-      "has_pages": false,
-      "has_discussions": false,
-      "forks_count": 0,
-      "mirror_url": null,
-      "archived": false,
-      "disabled": false,
-      "open_issues_count": 0,
-      "license": null,
-      "allow_forking": true,
-      "is_template": false,
-      "web_commit_signoff_required": false,
-      "topics": [
-  
-      ],
-      "visibility": "public",
-      "forks": 0,
-      "open_issues": 0,
-      "watchers": 0,
-      "default_branch": "main"
-    },
-    "sender": {
-      "login": "gravitational-member",
-      "id": 29820606,
-      "node_id": "MDQ6VXNlcjI5ODIwNjA2",
-      "avatar_url": "https://avatars.githubusercontent.com/u/29820606?v=4",
-      "gravatar_id": "",
-      "url": "https://api.github.com/users/gravitational-member",
-      "html_url": "https://github.com/gravitational-member",
-      "followers_url": "https://api.github.com/users/gravitational-member/followers",
-      "following_url": "https://api.github.com/users/gravitational-member/following{/other_user}",
-      "gists_url": "https://api.github.com/users/gravitational-member/gists{/gist_id}",
-      "starred_url": "https://api.github.com/users/gravitational-member/starred{/owner}{/repo}",
-      "subscriptions_url": "https://api.github.com/users/gravitational-member/subscriptions",
-      "organizations_url": "https://api.github.com/users/gravitational-member/orgs",
-      "repos_url": "https://api.github.com/users/gravitational-member/repos",
-      "events_url": "https://api.github.com/users/gravitational-member/events{/privacy}",
-      "received_events_url": "https://api.github.com/users/gravitational-member/received_events",
-      "type": "User",
-      "user_view_type": "public",
-      "site_admin": false
-    },
-    "installation": {
-      "id": 62588515,
-      "node_id": "MDIzOkludGVncmF0aW9uSW5zdGFsbGF0aW9uNjI1ODg1MTU="
-    }
-  }
