+++ tools/approval-service/internal/eventprocessor/dispatcher.go
@@ -8,17 +8,17 @@ import (
 	"log/slog"
 	"time"
 
-	"github.com/gravitational/shared-workflows/tools/approval-service/internal/approvalservice/config"
-	"github.com/gravitational/shared-workflows/tools/approval-service/internal/approvalservice/coordination"
-	"github.com/gravitational/shared-workflows/tools/approval-service/internal/approvalservice/eventprocessor/store"
-	"github.com/gravitational/shared-workflows/tools/approval-service/internal/approvalservice/sources/githubevents"
+	"github.com/gravitational/shared-workflows/tools/approval-service/internal/config"
+	"github.com/gravitational/shared-workflows/tools/approval-service/internal/coordination"
+	"github.com/gravitational/shared-workflows/tools/approval-service/internal/eventprocessor/store"
+	"github.com/gravitational/shared-workflows/tools/approval-service/internal/eventsources/githubevents"
 	"github.com/gravitational/teleport/api/types"
 	"golang.org/x/sync/errgroup"
 )
 
-// Processor manages changes to the state of Access Requests and their relation to deployment events (e.g. GitHub deployment review events).
+// Dispatcher manages changes to the state of Access Requests and their relation to deployment events (e.g. GitHub deployment review events).
 // It acts as a broker for the various event sources, ensuring that events are processed by the appropriate consumers.
-type Processor struct {
+type Dispatcher struct {
 	store store.ProcessorService
 
 	teleportUser    string
@@ -43,23 +43,23 @@ type Consumers struct {
 	GitHub []*GitHubConsumer
 }
 
-// Opt is a functional option for configuring the Processor.
-type Opt func(p *Processor) error
+// Opt is a functional option for configuring the Dispatcher.
+type Opt func(d *Dispatcher) error
 
-// WithLogger sets the logger for the Processor.
+// WithLogger sets the logger for the Dispatcher.
 func WithLogger(logger *slog.Logger) Opt {
-	return func(p *Processor) error {
+	return func(d *Dispatcher) error {
 		if logger == nil {
 			return errors.New("logger cannot be nil")
 		}
-		p.log = logger
+		d.log = logger
 		return nil
 	}
 }
 
-// New creates a new Processor instance.
-func New(ctx context.Context, teleConfig config.Teleport, sp *Consumers, coordinator coordination.Coordinator, opts ...Opt) (*Processor, error) {
-	p := &Processor{
+// NewDispatcher creates a new Dispatcher instance.
+func NewDispatcher(ctx context.Context, teleConfig config.Teleport, sp *Consumers, coordinator coordination.Coordinator, opts ...Opt) (*Dispatcher, error) {
+	d := &Dispatcher{
 		teleportUser:    teleConfig.User,
 		requestTTLHours: cmp.Or(time.Duration(teleConfig.RequestTTLHours)*time.Hour, 7*24*time.Hour),
 		githubConsumers: make(map[string]*GitHubConsumer),
@@ -68,32 +68,32 @@ func New(ctx context.Context, teleConfig config.Teleport, sp *Consumers, coordin
 	}
 
 	for _, opt := range opts {
-		if err := opt(p); err != nil {
+		if err := opt(d); err != nil {
 			return nil, fmt.Errorf("applying option: %w", err)
 		}
 	}
 
 	for _, gh := range sp.GitHub {
-		p.log.Info("Registering GitHub consumers", "id", githubID(gh.Org, gh.Repo))
-		p.githubConsumers[githubID(gh.Org, gh.Repo)] = gh
+		d.log.Info("Registering GitHub consumers", "id", githubID(gh.Org, gh.Repo))
+		d.githubConsumers[githubID(gh.Org, gh.Repo)] = gh
 	}
 
-	return p, nil
+	return d, nil
 }
 
-// Run starts the Processor and starts receiving events from the event sources.
-func (p *Processor) Run(ctx context.Context) error {
+// Run starts the Dispatcher and starts receiving events from the event sources.
+func (d *Dispatcher) Run(ctx context.Context) error {
 	eg, ctx := errgroup.WithContext(ctx)
 
 	eg.Go(func() error {
-		return p.githubEventListener(ctx)
+		return d.githubEventListener(ctx)
 	})
 
 	return eg.Wait()
 }
 
 // HandleReview will handle updates to the state of a Teleport Access Request.
-func (p *Processor) HandleReview(ctx context.Context, req types.AccessRequest) error {
+func (p *Dispatcher) HandleReview(ctx context.Context, req types.AccessRequest) error {
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
